(GH-8) Allow referencing tags in target_mapping
Previously, the `tags` attribute couldn't be effectively used in
target_mapping because it contains an array of maps with "key" and
"value" fields. That makes it impossible to find the value of a specific
desired tag.

We now convert the `tags` attribute into a standard key/value map so
that tags can be looked up using `tags.Name` syntax.