Add aws_access_key_id and aws_secret_access_key parameters
Previously, the only ways to provide credentials to the
resolve_reference task were to put the credentials in a file or in an
environment variable. Both of those require managing some aspect of the
environment external to the task. To avoid that requirement, the task
now also accepts these credentials directly as parameters. If both
a `credentials` file and one of the individual credential parameters are
supplied, the contents of the credentials file will take precedence.